refactor: Rename tsWorker.ts to CustomTypeScriptWorker.ts
